Remove "Edit this Page" link
This was surprisingly tricky, mostly because the docs are a bit opaque.
To make this work we need to set the 'edit_url' content: property in the
playbook, AND to set the CI environment variable to 'true' (even when
running locally). See

https://antora.zulipchat.com/#narrow/stream/282400-users/topic/Can't.20hide.20'Edit.20this.20Page'.20Link

I modified the Makefile which can be used for local builds accordingly
and hopefully, Actions CI will set the variable. If not I'll modify the
ci.yml as well.